+++ crates/client/tests/app.rs
@@ -447,7 +447,7 @@ async fn dump_network_stats() {
 
     let network_stats = app_ws.dump_network_stats().await.unwrap();
 
-    assert_eq!("kitsune2-core-mem", network_stats.backend);
+    assert_eq!("BackendLibDataChannel", network_stats.backend);
 }
 
 #[tokio::test(flavor = "multi_thread")]