+++ crates/holochain/CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+- Feat: Upgrade Kitsune2 to v0.3.2 \#5449
 - Fix panic if a DNA bundle is provided to Holochain with missing resources. #5446
 
 ## 0.6.0-rc.1